# feat(performance): added missing indexes to `schema_log` table