--- a/packages/migrations/src/run-pg-migrations.ts
+++ b/packages/migrations/src/run-pg-migrations.ts
@@ -67,7 +67,6 @@ import migration_2024_06_11T10_10_00_ms_teams_webhook from './actions/2024.06.11
 import migration_2024_07_16T13_44_00_oidc_only_access from './actions/2024.07.16T13-44-00.oidc-only-access';
 import migration_2024_07_17T00_00_00_app_deployments from './actions/2024.07.17T00-00-00.app-deployments';
 import migration_2024_07_23T_09_36_00_schema_cleanup_tracker from './actions/2024.07.23T09.36.00.schema-cleanup-tracker';
-import migration_2024_12_22_improve_version_index from './actions/2024.12.22T00-00-00.improve-version-index';
 import { runMigrations } from './pg-migrator';
 
 export const runPGMigrations = async (args: { slonik: DatabasePool; runTo?: string }) =>
@@ -148,6 +147,6 @@ export const runPGMigrations = async (args: { slonik: DatabasePool; runTo?: stri
       await import('./actions/2024.11.12T00-00-00.supertokens-9.1'),
       await import('./actions/2024.11.12T00-00-00.supertokens-9.2'),
       await import('./actions/2024.11.12T00-00-00.supertokens-9.3'),
-      migration_2024_12_22_improve_version_index,
+      await import('./actions/2024.12.23T00-00-00.improve-version-index'),
     ],
   });