This change adds two indexes to improve 2 db calls we are making frequently. 

## TL;DR 

Tested on a prod database clone.

- `idx_schema_log_action_created` - it took 42s to build, index size is 21 MB. reduced cost of running the same SELECT is 58k cost to 7k cost.
- `idx_schema_log_action_service` - took 3s to build, index size 5MB, cost reduced ~80%.

## `getSchemasOfVersion`

Seems like this one is executed more frequently and has major impact on performance (can take ~15s). 

```sql
EXPLAIN SELECT
            sl.id,
            sl.commit,
            sl.author,
            sl.action,
            sl.sdl,
            sl.created_at,
            sl.project_id,
            lower(sl.service_name) as service_name,
            sl.service_url,
            sl.metadata,
            sl.target_id,
            p.type
          FROM schema_version_to_log AS svl
          LEFT JOIN schema_log AS sl ON (sl.id = svl.action_id)
          LEFT JOIN projects as p ON (p.id = sl.project_id)
          WHERE
            svl.version_id = 'VERSION_ID'
            AND sl.action = 'PUSH'
            AND p.type != 'CUSTOM'
          ORDER BY
            sl.created_at DESC
```

Here's an example `EXPLAIN` plan for this one:

```
Gather Merge  (cost=59371.70..59397.14 rows=218 width=398)
  Workers Planned: 2
  ->  Sort  (cost=58371.68..58371.95 rows=109 width=398)
        Sort Key: sl.created_at DESC
        ->  Nested Loop  (cost=0.83..58367.99 rows=109 width=398)
              ->  Nested Loop  (cost=0.56..58333.45 rows=109 width=280)
                    ->  Parallel Seq Scan on schema_log sl  (cost=0.00..56531.26 rows=2697 width=280)
                          Filter: (action = 'PUSH'::text)
                    ->  Index Only Scan using schema_version_to_log_pkey on schema_version_to_log svl  (cost=0.56..28.91 rows=262 width=16)
                          Index Cond: ((version_id = 'x'::uuid) AND (action_id = sl.id))
              ->  Index Scan using projects_pkey on projects p  (cost=0.27..0.31 rows=1 width=134)
                    Index Cond: (id = sl.project_id)
                    Filter: ((type)::text <> 'CUSTOM'::text)
```

The `Parallel Seq Scan on schema_log sl  (cost=0.00..56531.26 rows=2697 width=280)` line is the heaviest, and it basically looks for `action='PUSH'` and we don't have an index for that kind of query. I added an index that also deals with the sorting (`idx_schema_log_action_created`).

## `getServiceSchemaOfVersion`

```sql
SELECT
            sl.id,
            sl.commit,
            sl.author,
            sl.action,
            sl.sdl,
            sl.created_at,
            sl.project_id,
            lower(sl.service_name) as service_name,
            sl.service_url,
            sl.target_id,
            p.type
          FROM schema_version_to_log AS svl
          LEFT JOIN schema_log AS sl ON (sl.id = svl.action_id)
          LEFT JOIN projects as p ON (p.id = sl.project_id)
          WHERE
            svl.version_id = 'x'
            AND sl.action = 'PUSH'
            AND p.type != 'CUSTOM'
            AND lower(sl.service_name) = lower('g')
```

Here's the query plan:

```
Nested Loop  (cost=1002.91..60289.80 rows=1 width=366)
  ->  Gather  (cost=1002.64..60286.75 rows=1 width=248)
        Workers Planned: 2
        ->  Nested Loop  (cost=2.64..59286.65 rows=1 width=248)
              ->  Parallel Seq Scan on schema_log sl  (cost=0.00..59228.16 rows=13 width=248)
                    Filter: ((action = 'PUSH'::text) AND (lower(service_name) = 'g'::text))
              ->  Bitmap Heap Scan on schema_version_to_log svl  (cost=2.64..490.41 rows=262 width=16)
                    Recheck Cond: ((version_id = 'x'::uuid) AND (action_id = sl.id))
                    ->  Bitmap Index Scan on schema_version_to_log_pkey  (cost=0.00..2.57 rows=262 width=0)
                          Index Cond: ((version_id = 'x'::uuid) AND (action_id = sl.id))
  ->  Index Scan using projects_pkey on projects p  (cost=0.27..3.04 rows=1 width=134)
        Index Cond: (id = sl.project_id)
        Filter: ((type)::text <> 'CUSTOM'::text)
```

Similar, but in this we also query by the service name, and we don't have sorting. I added `idx_schema_log_action_service`. 

This one took `4s` on production clone, and size is 5MB.